+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Support vanilla node.js http servers
+
 1.0.3 / 2014-05-20
 ==================